requirements.txt
disnake
python-dotenv
groq 
discord-handler
aiohttp 

